pyproject.toml
[tool.poetry]
name = "assignmentcura"
version = "0.1.0"
description = "Cura Assignment"
authors = ["Mithun"]
license = "BDS3"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
